+++ paddle/fluid/eager/to_static/run_program_func.cc
@@ -20,6 +20,7 @@
 #include "paddle/fluid/eager/to_static/run_program_utils.h"
 #include "paddle/fluid/eager/utils.h"
 #include "paddle/fluid/framework/tensor_ref_array.h"
+#include "paddle/phi/api/lib/data_transform.h"
 #include "paddle/phi/core/memory/allocation/allocator.h"
 #include "paddle/pir/include/core/block.h"
 #include "paddle/pir/include/core/builtin_type.h"
@@ -30,20 +31,24 @@ namespace egr::to_static {
 
 namespace {
 
-void clear_no_grad_edges_with_partial_block(
-    const std::vector<paddle::Tensor>& params,
-    const paddle::framework::BlockDesc* forward_block_desc,
-    const paddle::framework::BlockDesc* backward_block_desc,
-    egr::GradNodeBase* grad_node,
-    size_t slot_id) {
-  for (size_t i = 0; i < params.size(); ++i) {
-    auto p_grad_name = paddle::framework::GradVarName(params[i].name());
-    if (!forward_block_desc->HasVar(p_grad_name) &&
-        !backward_block_desc->HasVar(p_grad_name)) {
-      VLOG(3) << "clear edge of " << p_grad_name;
-      grad_node->MutableOutputMeta()[slot_id][i].GetMutableEdge().Clear();
+void clear_unused_out_var_in_backward(const std::vector<std::string>& out_names,
+                                      const pir::Block* backward_block,
+                                      paddle::framework::Scope* scope) {
+  std::deque<std::shared_ptr<paddle::memory::Allocation>>* garbages =
+      new std::deque<std::shared_ptr<paddle::memory::Allocation>>();
+  for (auto out_name : out_names) {
+    if (!backward_block->kwargs().count(out_name)) {
+      auto var = scope->FindVar(out_name);
+      if (var == nullptr) {
+        continue;
+      }
+      if (var->IsType<phi::DenseTensor>()) {
+        garbages->emplace_back(
+            var->GetMutable<phi::DenseTensor>()->MoveMemoryHolder());
+      }
     }
   }
+  delete garbages;
 }
 
 // Filter params without grads in global block. In this case, we will
@@ -63,64 +68,7 @@ void pir_clear_no_grad_edges(
   }
 }
 
-void clear_unused_out_var_in_backward(
-    const std::vector<paddle::Tensor*>& out,
-    const paddle::framework::BlockDesc* backward_block,
-    paddle::framework::Scope* scope) {
-  std::deque<std::shared_ptr<paddle::memory::Allocation>>* garbages =
-      new std::deque<std::shared_ptr<paddle::memory::Allocation>>();
-  for (auto* out_tensor : out) {
-    if (!backward_block->HasVar(out_tensor->name())) {
-      auto var = scope->FindVar(out_tensor->name());
-      if (var == nullptr) {
-        continue;
-      }
-      if (var->IsType<phi::DenseTensor>()) {
-        garbages->emplace_back(
-            var->GetMutable<phi::DenseTensor>()->MoveMemoryHolder());
-      }
-    }
-  }
-  delete garbages;
-}
-
-void pir_clear_unused_out_var_in_backward(
-    const std::vector<std::string>& out_names,
-    const pir::Block* backward_block,
-    paddle::framework::Scope* scope) {
-  std::deque<std::shared_ptr<paddle::memory::Allocation>>* garbages =
-      new std::deque<std::shared_ptr<paddle::memory::Allocation>>();
-  for (auto out_name : out_names) {
-    if (!backward_block->kwargs().count(out_name)) {
-      auto var = scope->FindVar(out_name);
-      if (var == nullptr) {
-        continue;
-      }
-      if (var->IsType<phi::DenseTensor>()) {
-        garbages->emplace_back(
-            var->GetMutable<phi::DenseTensor>()->MoveMemoryHolder());
-      }
-    }
-  }
-  delete garbages;
-}
-
 std::vector<paddle::Tensor> filter_unused_input_var_in_backward(
-    const std::vector<paddle::Tensor>& x,
-    const std::vector<std::string>& x_names,
-    const paddle::framework::BlockDesc* backward_block) {
-  auto filter_x = std::vector<paddle::Tensor>(x);
-  for (size_t i = 0; i < x.size(); i++) {
-    if (!backward_block->HasVar(x_names[i])) {
-      auto fake = paddle::Tensor(std::make_shared<phi::DenseTensor>());
-      fake.set_name(paddle::framework::kFakeVarName);
-      filter_x[i] = fake;
-    }
-  }
-  return filter_x;
-}
-
-std::vector<paddle::Tensor> pir_filter_unused_input_var_in_backward(
     const std::vector<paddle::Tensor>& x,
     const paddle::framework::AttributeMap& attrs) {
   const auto& names =
@@ -136,7 +84,7 @@ std::vector<paddle::Tensor> pir_filter_unused_input_var_in_backward(
   return filter_x;
 }
 
-std::vector<paddle::Tensor> pir_filter_no_need_buffer_input_var_in_backward(
+std::vector<paddle::Tensor> filter_no_need_buffer_input_var_in_backward(
     const std::vector<paddle::Tensor>& x,
     const paddle::framework::AttributeMap& attrs) {
   const auto& forward_inputs_names =
@@ -186,6 +134,59 @@ std::vector<paddle::Tensor> Trans2ContiguousTensors(
   return res;
 }
 
+// TODO(cleanup-legacy-ir): Cleanup these legacy functions
+void legacy_clear_no_grad_edges_with_partial_block(
+    const std::vector<paddle::Tensor>& params,
+    const paddle::framework::BlockDesc* forward_block_desc,
+    const paddle::framework::BlockDesc* backward_block_desc,
+    egr::GradNodeBase* grad_node,
+    size_t slot_id) {
+  for (size_t i = 0; i < params.size(); ++i) {
+    auto p_grad_name = paddle::framework::GradVarName(params[i].name());
+    if (!forward_block_desc->HasVar(p_grad_name) &&
+        !backward_block_desc->HasVar(p_grad_name)) {
+      VLOG(3) << "clear edge of " << p_grad_name;
+      grad_node->MutableOutputMeta()[slot_id][i].GetMutableEdge().Clear();
+    }
+  }
+}
+
+void legacy_clear_unused_out_var_in_backward(
+    const std::vector<paddle::Tensor*>& out,
+    const paddle::framework::BlockDesc* backward_block,
+    paddle::framework::Scope* scope) {
+  std::deque<std::shared_ptr<paddle::memory::Allocation>>* garbages =
+      new std::deque<std::shared_ptr<paddle::memory::Allocation>>();
+  for (auto* out_tensor : out) {
+    if (!backward_block->HasVar(out_tensor->name())) {
+      auto var = scope->FindVar(out_tensor->name());
+      if (var == nullptr) {
+        continue;
+      }
+      if (var->IsType<phi::DenseTensor>()) {
+        garbages->emplace_back(
+            var->GetMutable<phi::DenseTensor>()->MoveMemoryHolder());
+      }
+    }
+  }
+  delete garbages;
+}
+
+std::vector<paddle::Tensor> legacy_filter_unused_input_var_in_backward(
+    const std::vector<paddle::Tensor>& x,
+    const std::vector<std::string>& x_names,
+    const paddle::framework::BlockDesc* backward_block) {
+  auto filter_x = std::vector<paddle::Tensor>(x);
+  for (size_t i = 0; i < x.size(); i++) {
+    if (!backward_block->HasVar(x_names[i])) {
+      auto fake = paddle::Tensor(std::make_shared<phi::DenseTensor>());
+      fake.set_name(paddle::framework::kFakeVarName);
+      filter_x[i] = fake;
+    }
+  }
+  return filter_x;
+}
+
 }  // namespace
 
 void run_program_ad_func(
@@ -195,98 +196,8 @@ void run_program_ad_func(
     std::vector<paddle::framework::Scope*>& step_scope,  // NOLINT
     const paddle::framework::AttributeMap& attrs) {
   // Prepare Autograd Meta
-  VLOG(2) << "start run run_program ad function.";
-  auto deref_out = details::DereferenceTensors(out);
-  std::vector<egr::AutogradMeta*> p_autograd_x =
-      egr::EagerUtils::nullable_autograd_meta(x);
-  std::vector<egr::AutogradMeta*> p_autograd_params =
-      egr::EagerUtils::nullable_autograd_meta(params);
-  std::vector<egr::AutogradMeta*> p_autograd_outs =
-      egr::EagerUtils::nullable_autograd_meta(deref_out);
-
-  bool trace_backward = egr::Controller::Instance().HasGrad();
-  bool require_any_grad = egr::EagerUtils::ComputeRequireGrad(
-      trace_backward, &p_autograd_x, &p_autograd_params);
-
-  VLOG(2) << "start run run_program with require_any_grad = "
-          << require_any_grad;
-  auto x_tmp = Trans2ContiguousTensors(x);
-  auto params_tmp = Trans2ContiguousTensors(params);
-  // Call forward function
-  // if require_any_grad is False, don't save any middle vars.
-  int64_t place_hash_key = 0;
-  for (const paddle::Tensor& tensor : x) {
-    int64_t device_type = static_cast<int64_t>(tensor.place().GetType());
-    place_hash_key = hash_with_seed(place_hash_key, device_type);
-  }
-  RunProgramAPI(x_tmp,
-                params_tmp,
-                out,
-                step_scope,
-                require_any_grad,
-                attrs,
-                place_hash_key);
-  VLOG(2) << "start run run_program grad";
-  auto is_test = false;
-  if (attrs.count("is_test")) {
-    is_test = PADDLE_GET_CONST(bool, attrs.at("is_test"));
-  }
-  if (!is_test && require_any_grad) {
-    const auto& x_names =
-        PADDLE_GET_CONST(std::vector<std::string>, attrs.at("x_names"));
-
-    // Create GradOpNode (1 means [out_grad], 2 means [x_grad, paramx_grad])
-    auto grad_node = std::make_shared<GradNodeRunProgram>(1, 2);
-
-    // Set place hash keys for backward
-    grad_node->SetPlaceHashKey(place_hash_key);
-
-    // Set Attributes
-    grad_node->SetAttrMap(attrs);
-
-    auto* forward_global_block = PADDLE_GET_CONST(
-        paddle::framework::BlockDesc*, attrs.at("forward_global_block"));
-    auto* backward_global_block = PADDLE_GET_CONST(
-        paddle::framework::BlockDesc*, attrs.at("backward_global_block"));
-    // Clear unused x vars
-    auto filter_x = filter_unused_input_var_in_backward(
-        x_tmp, x_names, backward_global_block);
-    // Set TensorWrappers
-    grad_node->SetFwdX(filter_x);
-    // Clear unused out vars
-    clear_unused_out_var_in_backward(out, backward_global_block, step_scope[0]);
-
-    grad_node->SetFwdParams(params_tmp);
-    grad_node->SetStepScope(step_scope);
-
-    grad_node->SetGradOutMeta(x, /*slot id*/ 0);
-    grad_node->SetGradOutMeta(params, /*slot id*/ 1);
-
-    VLOG(2) << "clear_no_grad_edges.";
-    clear_no_grad_edges_with_partial_block(params,
-                                           forward_global_block,
-                                           backward_global_block,
-                                           grad_node.get(),
-                                           /*slot id*/ 1);
-
-    grad_node->SetGradInMeta(deref_out, 0);
-
-    egr::EagerUtils::SetOutRankWithSlot(&p_autograd_outs, 0);
-
-    // Set History for output set current Grad Node for
-    egr::EagerUtils::SetHistory(&p_autograd_outs, grad_node);
-  }
-}
-
-void pir_run_program_ad_func(
-    const std::vector<paddle::Tensor>& x,
-    const std::vector<paddle::Tensor>& params,
-    std::vector<paddle::Tensor*>& out,                   // NOLINT
-    std::vector<paddle::framework::Scope*>& step_scope,  // NOLINT
-    const paddle::framework::AttributeMap& attrs) {
-  // Prepare Autograd Meta
   VLOG(2) << "start run pir run_program ad function.";
-  auto deref_out = details::DereferenceTensors(out);
+  auto deref_out = egr::to_static::DereferenceTensors(out);
   std::vector<egr::AutogradMeta*> p_autograd_x =
       egr::EagerUtils::nullable_autograd_meta(x);
   std::vector<egr::AutogradMeta*> p_autograd_params =
@@ -313,16 +224,16 @@ void pir_run_program_ad_func(
     int64_t device_type = static_cast<int64_t>(tensor.place().GetType());
     place_hash_key = hash_with_seed(place_hash_key, device_type);
   }
-  PirRunProgramAPI(x_tmp,
-                   params_tmp,
-                   out,
-                   step_scope,
-                   require_any_grad,
-                   attrs,
-                   place_hash_key);
+  egr::to_static::RunProgramImpl(x_tmp,
+                                 params_tmp,
+                                 out,
+                                 step_scope,
+                                 require_any_grad,
+                                 attrs,
+                                 place_hash_key);
   if (!is_test && require_any_grad) {
     // Create GradOpNode (1 means [out_grad], 2 means [x_grad, paramx_grad])
-    auto grad_node = std::make_shared<PirGradNodeRunProgram>(1, 2);
+    auto grad_node = std::make_shared<GradNodeRunProgram>(1, 2);
 
     // Set place hash keys for backward
     grad_node->SetPlaceHashKey(place_hash_key);
@@ -337,8 +248,8 @@ void pir_run_program_ad_func(
     // We need to filter both of them.
     // For the first kind, we can create a empty Tensor to replace it.
     // For the second kind, we need to keep the meta only Tensor.
-    auto filter_x = pir_filter_no_need_buffer_input_var_in_backward(
-        pir_filter_unused_input_var_in_backward(x_tmp, attrs), attrs);
+    auto filter_x = filter_no_need_buffer_input_var_in_backward(
+        filter_unused_input_var_in_backward(x_tmp, attrs), attrs);
     // Set TensorWrappers
     grad_node->SetFwdX(filter_x);
 
@@ -349,7 +260,7 @@ void pir_run_program_ad_func(
     const auto& backward_params_grad_names =
         PADDLE_GET_CONST(std::vector<std::string>, attrs.at("bp_g_names"));
 
-    pir_clear_unused_out_var_in_backward(
+    clear_unused_out_var_in_backward(
         forward_outputs_names, backward_program->block(), step_scope[0]);
 
     grad_node->SetFwdParams(params_tmp);
@@ -366,6 +277,97 @@ void pir_run_program_ad_func(
                             backward_program->block(),
                             grad_node.get(),
                             /*slot id*/ 1);
+
+    grad_node->SetGradInMeta(deref_out, 0);
+
+    egr::EagerUtils::SetOutRankWithSlot(&p_autograd_outs, 0);
+
+    // Set History for output set current Grad Node for
+    egr::EagerUtils::SetHistory(&p_autograd_outs, grad_node);
+  }
+}
+
+void legacy_run_program_ad_func(
+    const std::vector<paddle::Tensor>& x,
+    const std::vector<paddle::Tensor>& params,
+    std::vector<paddle::Tensor*>& out,                   // NOLINT
+    std::vector<paddle::framework::Scope*>& step_scope,  // NOLINT
+    const paddle::framework::AttributeMap& attrs) {
+  // Prepare Autograd Meta
+  VLOG(2) << "start run run_program ad function.";
+  auto deref_out = egr::to_static::DereferenceTensors(out);
+  std::vector<egr::AutogradMeta*> p_autograd_x =
+      egr::EagerUtils::nullable_autograd_meta(x);
+  std::vector<egr::AutogradMeta*> p_autograd_params =
+      egr::EagerUtils::nullable_autograd_meta(params);
+  std::vector<egr::AutogradMeta*> p_autograd_outs =
+      egr::EagerUtils::nullable_autograd_meta(deref_out);
+
+  bool trace_backward = egr::Controller::Instance().HasGrad();
+  bool require_any_grad = egr::EagerUtils::ComputeRequireGrad(
+      trace_backward, &p_autograd_x, &p_autograd_params);
+
+  VLOG(2) << "start run run_program with require_any_grad = "
+          << require_any_grad;
+  auto x_tmp = Trans2ContiguousTensors(x);
+  auto params_tmp = Trans2ContiguousTensors(params);
+  // Call forward function
+  // if require_any_grad is False, don't save any middle vars.
+  int64_t place_hash_key = 0;
+  for (const paddle::Tensor& tensor : x) {
+    int64_t device_type = static_cast<int64_t>(tensor.place().GetType());
+    place_hash_key = hash_with_seed(place_hash_key, device_type);
+  }
+  egr::to_static::LegacyRunProgramImpl(x_tmp,
+                                       params_tmp,
+                                       out,
+                                       step_scope,
+                                       require_any_grad,
+                                       attrs,
+                                       place_hash_key);
+  VLOG(2) << "start run run_program grad";
+  auto is_test = false;
+  if (attrs.count("is_test")) {
+    is_test = PADDLE_GET_CONST(bool, attrs.at("is_test"));
+  }
+  if (!is_test && require_any_grad) {
+    const auto& x_names =
+        PADDLE_GET_CONST(std::vector<std::string>, attrs.at("x_names"));
+
+    // Create GradOpNode (1 means [out_grad], 2 means [x_grad, paramx_grad])
+    auto grad_node = std::make_shared<GradNodeLegacyRunProgram>(1, 2);
+
+    // Set place hash keys for backward
+    grad_node->SetPlaceHashKey(place_hash_key);
+
+    // Set Attributes
+    grad_node->SetAttrMap(attrs);
+
+    auto* forward_global_block = PADDLE_GET_CONST(
+        paddle::framework::BlockDesc*, attrs.at("forward_global_block"));
+    auto* backward_global_block = PADDLE_GET_CONST(
+        paddle::framework::BlockDesc*, attrs.at("backward_global_block"));
+    // Clear unused x vars
+    auto filter_x = legacy_filter_unused_input_var_in_backward(
+        x_tmp, x_names, backward_global_block);
+    // Set TensorWrappers
+    grad_node->SetFwdX(filter_x);
+    // Clear unused out vars
+    legacy_clear_unused_out_var_in_backward(
+        out, backward_global_block, step_scope[0]);
+
+    grad_node->SetFwdParams(params_tmp);
+    grad_node->SetStepScope(step_scope);
+
+    grad_node->SetGradOutMeta(x, /*slot id*/ 0);
+    grad_node->SetGradOutMeta(params, /*slot id*/ 1);
+
+    VLOG(2) << "clear_no_grad_edges.";
+    legacy_clear_no_grad_edges_with_partial_block(params,
+                                                  forward_global_block,
+                                                  backward_global_block,
+                                                  grad_node.get(),
+                                                  /*slot id*/ 1);
 
     grad_node->SetGradInMeta(deref_out, 0);
