+++ paddle/fluid/eager/to_static/run_program_func.h
@@ -28,10 +28,11 @@ void run_program_ad_func(
     std::vector<paddle::framework::Scope*>& step_scope,  // NOLINT
     const paddle::framework::AttributeMap& attrs);
 
-void pir_run_program_ad_func(
+void legacy_run_program_ad_func(
     const std::vector<paddle::Tensor>& x,
     const std::vector<paddle::Tensor>& params,
     std::vector<paddle::Tensor*>& out,                   // NOLINT
     std::vector<paddle::framework::Scope*>& step_scope,  // NOLINT
     const paddle::framework::AttributeMap& attrs);
+
 }  // namespace egr::to_static
