+++ paddle/fluid/pybind/eager_custom_python_api.h
@@ -101,7 +101,7 @@ static PyObject *eager_api_run_program(PyObject *self,
 
     VLOG(6) << "Finish Pir ConstructAttrMapFromPyArgs";
     tstate = PyEval_SaveThread();
-    egr::to_static::pir_run_program_ad_func(X, Params, Out, OutScope, attrs);
+    egr::to_static::run_program_ad_func(X, Params, Out, OutScope, attrs);
     PyEval_RestoreThread(tstate);
     tstate = nullptr;
     Py_RETURN_NONE;
